requirements.txt
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
scikit-learn==1.3.2
ta==0.10.2
requests==2.31.0
beautifulsoup4==4.12.2
streamlit==1.29.0
dash==2.16.1
dash-bootstrap-components==1.5.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
xgboost==2.0.3
lightgbm==4.1.0
